python: upstream merge: v2025.14.0 (#9613)
Addresses https://github.com/posit-dev/positron/issues/9541 and
https://github.com/posit-dev/positron/issues/9419.

Merging in v2025.14.0 of the upstream python extension, plus monthly
maintenance, plus adding support for Python 3.14.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Added preview support for Python 3.14
https://github.com/posit-dev/positron/issues/9419

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->
@:web @:win

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->
It might be good to manually test that everything works with Python
3.14. I recommend downloading from
https://www.python.org/downloads/release/python-3140rc3/ instead of
using `uv` because they don't have the latest release candidate as of
this writing. Watch out for
https://github.com/posit-dev/positron/issues/9655 if you're testing.

---------

Co-authored-by: positron-bot[bot] <173392469+positron-bot[bot]@users.noreply.github.com>