Use CodeOSS version for extension update check (#9675)
Address #9533 

Re-enables using the CodeOSS version when checking for extension
updates. This would reduce the amount of extension updates by updating
extensions to a version that is supported by the pending Positron
update.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Use CodeOSS version when checking for extension updates


### QA Notes
Testing can be done against the `dailies` update channel. Those builds
are currently including `codeoss_version` in the update metadata.
Release builds will include it as well on the next build published to
the `releases` update channel.

For testing in dev, turn off auto-updates and set the update channel to
`dailies`. Auto-updates don't work in dev and it will disable the update
service if enabled.