Show output during reconnect (#9665)
This change moves some bits around to improve the experience of
reconnecting/attaching to a session that is in the middle of a
computation.

Before the change, you had to wait until the session was idle to see any
computational results. After the change, you can reattach to a session
that's busy and resume watching the computational results stream to your
console.

Note that some language features (most notably the LSP and Variables
pane) are still unavailable until the kernel is idle again. The console
will remain in a _Reconnecting_ state to signal this.

The basic idea is to return right away from the `restoreSession` call
instead of waiting for the kernel to be ready. This call formerly waited
for the kernel to be ready so it could request the kernel info to return
to the client; this kernel info is now cached by Kallichore (note
version bump), so it can be returned without waiting.

The most fiddly part of this change is in the console, which previously
assumed that every time a session was attached, it was ready to run code
(formerly a safe assumption). Now we need to be more careful and only
mark the console as ready when we actually get a ready signal from the
supervisor.

Addresses https://github.com/posit-dev/positron/issues/8388.

#### New Features

- Posit Workbench: show results of in-progress computations when
reconnecting to a session (#8388)

#### Bug Fixes

- N/A


### QA Notes

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->