Data Explorer: Fix Python histogram inconsistencies for 0-row data frames (or 0 rows after filtering) (#9621)
Addresses #9527. https://github.com/posit-dev/ark/pull/932 adds fixes
for 1-row and 0-row data frames in R which addresses #9528.

I had noticed that histograms would fail to compute for pandas data
frames with zero rows. This normalizes and tests the behavior to make
sure it is consistent (and works) for both polars and pandas.