Include Quarto and Pandoc in REH builds (#9623)
This change adds Quarto (and, by extension, Pandoc) to the Remote
Extension Host builds of Positron, which include Remote SSH & Workbench.

Addresses https://github.com/posit-dev/positron/issues/9548.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Quarto and Pandoc are now included in the Remote SSH version of
Positron (#9548)

#### Bug Fixes

- N/A


### QA Notes

An easy way to see if Positron knows about its built-in copy of Pandoc
is to check the value of the `RSTUDIO_PANDOC` environment variable.