Assistant: Use more efficient format for language session variables (#9724)
This addresses #8992.

### Release Notes

This change makes it so that Positron Assistant uses a more efficient
format for sending session variables, which greatly reduces token usage.
It also limits the number of items sent to 400, to avoid pathologically
large cases.

The new format looks like this:

```
<variables description="Variables defined in the current session, in a pipe-delimited format, where each line is `name|kind|display_type`.">
x|number|int
y|number|int
hello_world|function|function
</variables>
```


### QA Notes

The main thing is to make sure that Positron Assistant continues to
behave the same as it did before, in terms of being able to understand
the current session context. So it should still be aware of what
variables are loaded, and what type they are. And this should be tested
across multiple LLMs, to make sure that the format can be understood by
them.

---------

Co-authored-by: Sam Clark <sam.clark@posit.co>
Co-authored-by: Jonathan McPherson <jonathan@posit.co>