fix notebook.workingDirectory setting for untitled notebooks (#9768)
Addresses #9631. Optimizes the handling of the schemes of URIs, so that
a bug is fixed where untitled notebooks weren't respecting the
`notebook.workingDirectory` setting.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fixed #9631


### QA Notes

@:notebooks @:web @:win

There's a reprex in the original issue.