Fix `positron.RuntimeCodeExecutionMode.Silent` behavior for queued code in the console (#9772)
- Addresses https://github.com/posit-dev/positron/issues/7205
- Closes https://github.com/posit-dev/positron/pull/9741

This is an internal, non-fork PR applying the commits from
https://github.com/posit-dev/positron/pull/9741, because of the problems
outlined in https://github.com/posit-dev/positron/issues/6628.

@:console @:interpreter @:sessions

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fix `positron.RuntimeCodeExecutionMode.Silent` behavior for queued
code in the console


### QA Notes

The motivating extension example in
https://github.com/posit-dev/positron/issues/7205 is a good example to
look at, and I also chose some existing tests to run here which run code
in the console from an extension (like the R package dev tests).

---------

Co-authored-by: Dianyi Yang <dianyi.yang@politics.ox.ac.uk>
Co-authored-by: kv9898 <105025148+kv9898@users.noreply.github.com>