More quoting and normalization around R binpaths, for Windows (#9780)
Addresses #9216 

My initial hunch was that this had to do with path separators, which
lead to me adding path normalization in many places. I still think this
is a good idea, so I've kept that.

But the real problem was that the R binpath wasn't *necessarily* quoted
when firing the `vscode.ShellExecution`s that back R package tasks. And
that is necessary for robustness on Windows.

It turns out, because most people have R installed below `C:\Program
Files`, that pesky space was causing the binpath to be quoted, which is
good and necessary. But if you have R installed somewhere else and you
cause it to be discovered by Positron and that somewhere else *doesn't*
have spaces or other challenging characters, the binpath won't be
quoted. And the task falls over.

Here's the quoting for R installed in the usual place (note the single
quotes around the binpath):

```
*  Executing task: 'C:\Program Files\R\R-4.5.1\bin\x64\R.exe' --quiet --no-restore --no-save -e 'devtools::check()' 

> devtools::check()
══ Documenting ═════════════════════════════════════════════════════════════════════════════════
ℹ Updating stringr documentation
ℹ Loading stringr
...  blah blah
```

Here's the **lack** of quoting for R installed in a more creative place:

```
Executing task: C:\quirkyRLocation\R-4.3.3\bin\x64\R.exe --quiet --no-restore --no-save -e 'devtools::check()' 

bash: C:quirkyRLocationR-4.3.3binx64R.exe: command not found

 *  The terminal process "C:\Program Files\Git\bin\bash.exe '--login', '-i', '-c', 'C:\quirkyRLocation\R-4.3.3\bin\x64\R.exe --quiet --no-restore --no-save -e 'devtools::check()''" terminated with exit code: 127.
```

With this PR, we force the binpath to be quoted unconditionally, which
gets R package tasks working again:

```
Executing task: & 'C:\quirkyRLocation\R-4.3.3\bin\x64\R.exe' --quiet --no-restore --no-save -e 'devtools::check()' 

> devtools::check()
══ Documenting ═══════════════════════════════════════════════════════════════
ℹ Updating stringr documentation
... blah blah
```

### QA Notes

Must be on Windows.

You must have R installed to a path that contains no spaces or other
quote-worthy characters and ensure that Positron discovers it. I used a
[CRAN installer for R
4.3](https://cran.r-project.org/bin/windows/base/old/) and installed it
to `C:\quirkyRLocation\R-4.3.3`. I then informed Positron about that by
specifying a custom root folder:

```
...
"positron.r.customRootFolders": [
  "C:/quirkyRLocation"
]
...
```

Prior to this PR, checking or testing an R package will fail with the
error shown above and in #9216.

With this PR, those package tests will succeed. I tested this with
Positron using both Git Bash and PowerShell, i.e. I played with this
setting:

```
...
//"terminal.integrated.defaultProfile.windows": "Git Bash",
"terminal.integrated.defaultProfile.windows": "PowerShell",
...
```

`@:ark` `@:win`