Data Explorer: Fix value parsing issues with thousands separator when formatting histogram tooltips (#9812)
Addresses #9798. For values 1000 or higher, the histogram bins were
being formatted using the cell formatter in the backend, introducing a
thousands separator which would foul up `parseFloat` which was used
later to reformat the tooltip (or handle the conversion to whole numbers
for integers).

This now displays the bins as formatted by the backend for floating
point data

<img width="367" height="119" alt="image"
src="https://github.com/user-attachments/assets/05f678f7-e24d-46cd-afef-d644485afe21"
/>

Integers over 1000 are fixed, too

<img width="348" height="105" alt="image"
src="https://github.com/user-attachments/assets/36768222-afe8-4ac1-9cbe-b62151d04074"
/>

There are some other refinements as described in #9801 but since this
issue is critical it would be better to get this fix in ASAP and address
those in a follow up PR.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

@:data-explorer