e2e: notebook infra/cleanup (#9831)
### Summary

The objective of this PR was to update some of the e2e Positron Notebook
infra to make the tests more maintainable and reliable (fixed some
flakes). I also removed a quite a few hardcoded waits that weren't
necessary once the assertions were improved. I will be keeping an eye on
them to quickly address any upcoming flakes.

### QA Notes

@:positron-notebooks