fix open-remote-ssh whitespace (#9863)
Our hard fork of the open-remote-ssh extension has a nice mixture of
files with tabs and files with spaces. This PR changes everything to
tabs, in order to align with our top-level
[settings.json](https://github.com/posit-dev/positron/blob/main/.vscode/settings.json#L6)
file, so that I don't have to keep worrying about temporarily disabling
formatting every time I need to work in this extension.