Assistant: Improve managing language models (#9882)
Ports #9799 from the 2025.10 branch into main.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

This merge resulted in a number of conflicts so worth double-checking
that things still work as intended!

---------

Co-authored-by: George Stagg <george.stagg@posit.co>
Co-authored-by: sharon wang <sharon-wang@users.noreply.github.com>
Co-authored-by: Tim Mok <timtmok@users.noreply.github.com>