Check `positron.assistant.enable` before turning on console Assistant actions (#9921)
This is a very minimal fix for
https://github.com/posit-dev/positron/issues/9919 that I think would not
be risky for the Workbench team to take in their next release.

A longer term, more broad fix for
https://github.com/posit-dev/positron/issues/9919 might involve some
complex set of configuration watchers in the Assistant `extension.ts`
that sort of turn off the `true` defaults if Assistant hasn't been
enabled. That would be fairly complex but we could more easily protect
ourselves against making this mistake again in the future.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Fixed the behavior of Assistant’s Console actions when Assistant is
not enabled at all.


### QA Notes

If you have a very fresh build of Positron with no state and you have
_not_ enabled Positron Assistant, then code that errors (for example,
`data[1]` in R) should return an error and traceback but _no_
Fix/Explain decorations.

---------

Co-authored-by: Sam Clark <sam.clark@posit.co>