Check `positron.assistant.enable` before turning on console Assistant actions (#9924)
This PR cherry picks #9921 to the `prerelease/2025.10` branch so that we
can do a patch release, if we decide we want to. This is a fairly low
risk, minimal change that I think the Workbench team could take in their
release.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Fixed the behavior of Assistant’s Console actions when Assistant is
not enabled at all.


### QA Notes

If you have a very fresh build of Positron with no state and you have
_not_ enabled Positron Assistant, then code that errors (for example,
`data[1]` in R) should return an error and traceback but _no_
Fix/Explain decorations.

---------

Co-authored-by: Sam Clark <sam.clark@posit.co>