Fix empty editor bug (#9996)
Addresses #9849 and #10011

The editor-dissapearing bug was caused by improper storage of the react
renderer in the notebook editor. Because of this when navigating back to
the notebook the editor would not realize it had rendered before and
restart everything which ran into bugs with things like disposed objects
etc.

This PR makes sure that we actually store the react renderer to prevent
premature cleanup.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

You should be able to open a notebook then navigate away from the tab
(make it not visible at all) and back and the notebook should still show
up.