update data explorer column selector to use new search (#10004)
Addresses #9847

Updates the column selector modal popup (which is a data grid under the
hood) to use the new backend search functionality we added to the data
explorer.

This should address the issue we were seeing where column names weren't
rendering in the drop down when the number of columns was greater than a
few hundred. Instead of trying to fetch and render all columns, only the
columns that are visible are shown and cached. When the user scrolls or
searches, we go out and fetch the column data that should be in view.
This is the same behavior we have in the other data grids, like the
summary panel.

Note: The RPC timeout issue described in #9846 is now also present in
the column selector modal popup when searching if there are a large
number of columns present. Searching will silently fail if it takes
longer than 5 seconds. Depsite this, the behavior is much improved than
what it was previously.

For keyboard functionality testing, I want to point out the focus and
keyboard behavior when using the search in the filter column selector
dropdown is slightly different than a traditional dropdown because of
some of the custom cursor selection behavior we have. The biggest thing
to be aware of is: the focus/selected styling in the dropdown is always
visible (and set to the first item in the dropdown) when a search is
active. This is a bit odd when searching because the _actual_ focus
indicator is in the search box. Pressing `Arrow Down` or `Tab` will move
the actual focus indicator into the dropdown but its not obvious that is
where focus went since the selected/focused state for the dropdown is
always present and there is no difference between focus and selection.
Everything works, its just a bit odd visually. This is how its always
worked and hasn't been changed in this PR! Just be aware when testing.

Here's an example of an accessible combobox pattern in case you are
curious what the differences are:
https://www.w3.org/WAI/ARIA/apg/patterns/combobox/examples/combobox-autocomplete-both/

**BEFORE - 5 million columns**


https://github.com/user-attachments/assets/db44403c-a0b6-43d1-9c0b-93da5fe51d51

**AFTER - Dropdown without search (6 columns)**


https://github.com/user-attachments/assets/aee88785-06bd-4cc1-8a71-f2766c5644cc

**AFTER - Dropdown with search (1000 columns)**


https://github.com/user-attachments/assets/4b323f7e-03bb-4783-b92c-b2e20a2a2d12

**AFTER - Dropdown with search (5 million columns)**


https://github.com/user-attachments/assets/6dfde728-1efc-42f7-8814-75fb15e845c9



### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Update data explorer column selector to use new search (#9847)

#### Bug Fixes

- N/A


### QA Notes
@:data-explorer
@:web
@:win
<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->