python: windows handling for undetermined versions (#10057)
There was some handling that created fake paths by looking for `./.venv`
and then immediately trying to start up `./.venv/bin/python`. The `js`
locator is just smarter about resolving this correctly, which is why we
do not see this problem there.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- #9186 Fixes bug where Positron created environments show as version
0.0.1.


### QA Notes

make a new project through new project wizard. should not get 0.0.1
versioned python