Configurable startup timeout for supervisor; better exit detection (#10042)
This change does not fix any known problems, but is intended to help us
get better diagnostics on issues like #8000 wherein the supervisor
appears to be timing out at startup. Specifically, it seems possible
that the supervisor is actually exited in these cases or is taking extra
long to start, so:

- make the timeout configurable so that if there actually is a situation
in which it takes a super long time for first boot (AV scan? insanely
slow hardware? something else?) it is survivable
- move the close listener handler to _before_ we open the terminal, so
that if the terminal closes immediately after starting we'll catch it
- periodically while polling, check to see if the process is alive; it's
possible it is exiting but the terminal isn't closing or the close
listener isn't getting triggered