Merging in upstream vscode-server for admin enforced settings (#10082)
#9573 

Brings in the changes to support admin enforced settings. Setting
`POSITRON_ENFORCED_SETTINGS` before starting Positron (web or desktop)
with the JSON for the settings (matching the format in `settings.json`)
will freeze the settings. They cannot be changed after startup and
altering the environment variable will have no effect. If an admin
changes the enforced settings, it will only affect new instances.

The desktop support has no use case at the moment but provides the
ability for implementing a mechanism for admins to enforce settings on
desktop as well. This would allow support for Linux, more complex
settings (upstream only allows settings that are booleans, strings, and
numbers), and for settings that aren't specifically marked as a policy
(upstream only allows a short list of settings to be enforced).

The images below used this value for the environment variable:
`"{\"positron.assistant.enabledProviders\":
[\"amazon-bedrock\",\"openai-api\"], \"positron.assistant.enable\":
true}"` This locks Assistant to enabled and sets the enabled providers
to Bedrock and OpenAI (aside from the default supported Anthropic and
Copilot).

<img width="823" height="842" alt="image"
src="https://github.com/user-attachments/assets/114b031e-cf30-4f74-82a4-efc48181e7ba"
/>

The `settings.json` shows that `echo` was set but it is ignored and
overridden by the admin setting. Also, the incorrect `openai` value is
set there but the admin setting sets it correctly so it also shows as
available in the dialog.

<img width="489" height="82" alt="image"
src="https://github.com/user-attachments/assets/2b96cadf-5f40-4aff-801c-c54155f56855"
/>

The settings editor decorates a setting as managed by the organization.
It cannot be altered. This decoration comes from upstream. The behaviour
around preventing changes is based on the existing upstream policy code.

### Release Notes

#### New Features

- Admin enforced settings

#### Bug Fixes

- N/A


### QA Notes
Attempting to change an enforced setting may allow it to be changed but
verify that the setting is locked even though the value changed. Often,
the value reverts once the setting loses focus and the UI syncs back to
the state.

Co-authored-by: positron-bot[bot] <173392469+positron-bot[bot]@users.noreply.github.com>